+++ src/Nelknet.LibSQL.Data/Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 // Make internal members visible to the test assembly
 [assembly: InternalsVisibleTo("Nelknet.LibSQL.Tests")]
+[assembly: InternalsVisibleTo("Nelknet.LibSQL.Benchmarks")]